get one lock and one cache_get call in case of sync_writes
