+++ Cargo.toml
@@ -36,7 +36,7 @@ path = "cached_proc_macro_types"
 optional = true
 
 [dependencies.hashbrown]
-version = "0.12"
+version = "0.13"
 default-features = false
 features = ["raw", "inline-more"]
 
@@ -100,7 +100,7 @@ version = "1"
 version = "1"
 
 [dev-dependencies.serial_test]
-version = "0.7"
+version = "0.10"
 
 [workspace]
 members = ["cached_proc_macro","examples/wasm"]
