+++ cached_proc_macro/Cargo.toml
@@ -18,7 +18,8 @@ proc-macro = true
 
 [dependencies]
 quote = "1.0.6"
-darling = "0.13.0"
+darling = "0.14.2"
+proc-macro2 = "1.0.49"
 
 [dependencies.syn]
 version = "1.0.27"
