+++ cached_proc_macro/src/lib.rs
@@ -1,12 +1,13 @@
+mod helpers;
+
 use darling::FromMeta;
+use helpers::*;
 use proc_macro::TokenStream;
 use quote::quote;
-use std::ops::Deref;
-use syn::punctuated::Punctuated;
 use syn::spanned::Spanned;
 use syn::{
-    parse_macro_input, parse_quote, parse_str, AttributeArgs, Block, ExprClosure, FnArg,
-    GenericArgument, Ident, ItemFn, Pat, PathArguments, ReturnType, Type,
+    parse_macro_input, parse_str, AttributeArgs, Block, ExprClosure, GenericArgument, Ident,
+    ItemFn, PathArguments, ReturnType, Type,
 };
 
 #[derive(FromMeta)]
@@ -88,45 +89,8 @@ pub fn cached(args: TokenStream, input: TokenStream) -> TokenStream {
     let output = signature.output.clone();
     let asyncness = signature.asyncness;
 
-    // pull out the names and types of the function inputs
-    let input_tys = inputs
-        .iter()
-        .map(|input| match input {
-            FnArg::Receiver(_) => panic!("methods (functions taking 'self') are not supported"),
-            FnArg::Typed(pat_type) => pat_type.ty.clone(),
-        })
-        .collect::<Vec<Box<Type>>>();
-
-    let input_names = inputs
-        .iter()
-        .map(|input| match input {
-            FnArg::Receiver(_) => panic!("methods (functions taking 'self') are not supported"),
-            FnArg::Typed(pat_type) => {
-                // if you define arguments as mutable, e.g.
-                // #[cached]
-                // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-                //     a += 1;
-                //     b += 1;
-                //     (a, b)
-                // }
-                // then we need to strip off the `mut` keyword from the
-                // variable identifiers so we can refer to arguments `a` and `b`
-                // instead of `mut a` and `mut b`
-                match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                }
-            }
-        })
-        .collect::<Vec<Box<Pat>>>();
+    let input_tys = get_input_types(&inputs);
+    let input_names = get_input_names(&inputs);
 
     // pull out the output type
     let output_ty = match &output {
@@ -136,99 +100,29 @@ pub fn cached(args: TokenStream, input: TokenStream) -> TokenStream {
 
     let output_span = output_ty.span();
     let output_ts = TokenStream::from(output_ty.clone());
-    let output_parts = output_ts
-        .clone()
-        .into_iter()
-        .filter_map(|tt| match tt {
-            proc_macro::TokenTree::Ident(ident) => Some(ident.to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
+    let output_parts = get_output_parts(&output_ts);
     let output_string = output_parts.join("::");
     let output_type_display = output_ts.to_string().replace(' ', "");
 
-    // if `with_cached_flag = true`, then enforce that the return type
-    // is something wrapped in `Return`. Either `Return<T>` or the
-    // fully qualified `cached::Return<T>`
-    if args.with_cached_flag
-        && !output_string.contains("Return")
-        && !output_string.contains("cached::Return")
-    {
-        return syn::Error::new(
-            output_span,
-            format!(
-                "\nWhen specifying `with_cached_flag = true`, \
-                    the return type must be wrapped in `cached::Return<T>`. \n\
-                    The following return types are supported: \n\
-                    |    `cached::Return<T>`\n\
-                    |    `std::result::Result<cachedReturn<T>, E>`\n\
-                    |    `std::option::Option<cachedReturn<T>>`\n\
-                    Found type: {t}.",
-                t = output_type_display
-            ),
-        )
-        .to_compile_error()
-        .into();
+    if check_with_cache_flag(args.with_cached_flag, output_string) {
+        return with_cache_flag_error(output_span, output_type_display);
     }
 
-    // Find the type of the value to store.
-    // Normally it's the same as the return type of the functions, but
-    // for Options and Results it's the (first) inner type. So for
-    // Option<u32>, store u32, for Result<i32, String>, store i32, etc.
-    let cache_value_ty = match (&args.result, &args.option) {
-        (false, false) => output_ty,
-        (true, true) => panic!("the result and option attributes are mutually exclusive"),
-        _ => match output.clone() {
-            ReturnType::Default => {
-                panic!("function must return something for result or option attributes")
-            }
-            ReturnType::Type(_, ty) => {
-                if let Type::Path(typepath) = *ty {
-                    let segments = typepath.path.segments;
-                    if let PathArguments::AngleBracketed(brackets) =
-                        &segments.last().unwrap().arguments
-                    {
-                        let inner_ty = brackets.args.first().unwrap();
-                        quote! {#inner_ty}
-                    } else {
-                        panic!("function return type has no inner type")
-                    }
-                } else {
-                    panic!("function return type too complex")
-                }
-            }
-        },
-    };
+    let cache_value_ty = find_value_type(args.result, args.option, &output, output_ty);
 
     // make the cache identifier
     let cache_ident = match args.name {
-        Some(name) => Ident::new(&name, fn_ident.span()),
+        Some(ref name) => Ident::new(name, fn_ident.span()),
         None => Ident::new(&fn_ident.to_string().to_uppercase(), fn_ident.span()),
     };
 
-    // make the cache key type and block that converts the inputs into the key type
-    let (cache_key_ty, key_convert_block) = match (&args.key, &args.convert, &args.cache_type) {
-        (Some(key_str), Some(convert_str), _) => {
-            let cache_key_ty = parse_str::<Type>(key_str).expect("unable to parse cache key type");
-
-            let key_convert_block =
-                parse_str::<Block>(convert_str).expect("unable to parse key convert block");
-
-            (quote! {#cache_key_ty}, quote! {#key_convert_block})
-        }
-        (None, Some(convert_str), Some(_)) => {
-            let key_convert_block =
-                parse_str::<Block>(convert_str).expect("unable to parse key convert block");
-
-            (quote! {}, quote! {#key_convert_block})
-        }
-        (None, None, _) => (
-            quote! {(#(#input_tys),*)},
-            quote! {(#(#input_names.clone()),*)},
-        ),
-        (Some(_), None, _) => panic!("key requires convert to be set"),
-        (None, Some(_), None) => panic!("convert requires key or type to be set"),
-    };
+    let (cache_key_ty, key_convert_block) = make_cache_key_type(
+        &args.key,
+        &args.convert,
+        &args.cache_type,
+        input_tys,
+        &input_names,
+    );
 
     // make the cache type and create statement
     let (cache_ty, cache_create) = match (
@@ -324,119 +218,80 @@ pub fn cached(args: TokenStream, input: TokenStream) -> TokenStream {
         _ => panic!("the result and option attributes are mutually exclusive"),
     };
 
-    let do_set_return_block = if asyncness.is_some() {
-        if args.sync_writes {
-            quote! {
-                // try to get a write lock first
-                let mut cache = #cache_ident.lock().await;
-                if let Some(result) = cache.cache_get(&key) {
-                    #return_cache_block
-                }
+    let set_cache_and_return = quote! {
+        #set_cache_block
+        result
+    };
+    let lock;
+    let function_call;
+    let cache_type;
+    if asyncness.is_some() {
+        lock = quote! {
+            // try to get a lock first
+            let mut cache = #cache_ident.lock().await;
+        };
 
-                // run the function and cache the result
-                async fn inner(#inputs) #output #body;
-                let result = inner(#(#input_names),*).await;
-                #set_cache_block
-                result
-            }
-        } else {
-            quote! {
-                // run the function and cache the result
-                async fn inner(#inputs) #output #body;
-                let result = inner(#(#input_names),*).await;
-                let mut cache = #cache_ident.lock().await;
-                #set_cache_block
-                result
-            }
-        }
-    } else if args.sync_writes {
-        quote! {
-            // try to get a write lock first
+        function_call = quote! {
+            // run the function and cache the result
+            async fn inner(#inputs) #output #body;
+            let result = inner(#(#input_names),*).await;
+        };
+
+        cache_type = quote! {
+            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<::cached::async_sync::Mutex<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| ::cached::async_sync::Mutex::new(#cache_create));
+        };
+    } else {
+        lock = quote! {
+            // try to get a lock first
             let mut cache = #cache_ident.lock().unwrap();
+        };
+
+        function_call = quote! {
+            // run the function and cache the result
+            fn inner(#inputs) #output #body;
+            let result = inner(#(#input_names),*);
+        };
+
+        cache_type = quote! {
+            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<std::sync::Mutex<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| std::sync::Mutex::new(#cache_create));
+        };
+    }
+
+    let prime_do_set_return_block = quote! {
+        #lock
+        #function_call
+        #set_cache_and_return
+    };
+
+    let do_set_return_block = if args.sync_writes {
+        quote! {
+            #lock
             if let Some(result) = cache.cache_get(&key) {
                 #return_cache_block
             }
-
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            #set_cache_block
-            result
+            #function_call
+            #set_cache_and_return
         }
     } else {
         quote! {
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            let mut cache = #cache_ident.lock().unwrap();
-            #set_cache_block
-            result
+            {
+                #lock
+                if let Some(result) = cache.cache_get(&key) {
+                    #return_cache_block
+                }
+            }
+            #function_call
+            #lock
+            #set_cache_and_return
         }
     };
 
-    // if you define arguments as mutable, e.g.
-    // #[cached]
-    // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-    //     a += 1;
-    //     b += 1;
-    //     (a, b)
-    // }
-    // then we want the `mut` keywords present on the "inner" function
-    // that wraps your actual block of code.
-    // If the `mut`s are also on the outer method, then you'll
-    // get compiler warnings about your arguments not needing to be `mut`
-    // when they really do need to be.
-    let mut signature_no_muts = signature;
-    let mut sig_inputs = Punctuated::new();
-    for inp in &signature_no_muts.inputs {
-        let item = match inp {
-            FnArg::Receiver(_) => inp.clone(),
-            FnArg::Typed(pat_type) => {
-                let mut pt = pat_type.clone();
-                let pat = match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                };
-                pt.pat = pat;
-                FnArg::Typed(pt)
-            }
-        };
-        sig_inputs.push(item);
-    }
-    signature_no_muts.inputs = sig_inputs;
+    let signature_no_muts = get_mut_signature(signature);
 
     // create a signature for the cache-priming function
     let prime_fn_ident = Ident::new(&format!("{}_prime_cache", &fn_ident), fn_ident.span());
     let mut prime_sig = signature_no_muts.clone();
     prime_sig.ident = prime_fn_ident;
-
-    let prime_do_set_return_block = if asyncness.is_some() {
-        quote! {
-            // run the function and cache the result
-            async fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*).await;
-            let mut cache = #cache_ident.lock().await;
-            #set_cache_block
-            result
-        }
-    } else {
-        quote! {
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            let mut cache = #cache_ident.lock().unwrap();
-            #set_cache_block
-            result
-        }
-    };
 
     // make cached static, cached function and prime cached function doc comments
     let cache_ident_doc = format!("Cached static for the [`{}`] function.", fn_ident);
@@ -445,70 +300,27 @@ pub fn cached(args: TokenStream, input: TokenStream) -> TokenStream {
         "This is a cached function that uses the [`{}`] cached static.",
         cache_ident
     );
-    if attributes.iter().any(|attr| attr.path.is_ident("doc")) {
-        attributes.push(parse_quote! { #[doc = ""] });
-        attributes.push(parse_quote! { #[doc = "# Caching"] });
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    } else {
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    }
+    fill_in_attributes(&mut attributes, cache_fn_doc_extra);
 
     // put it all together
-    let expanded = if asyncness.is_some() {
-        quote! {
-            // Cached static
-            #[doc = #cache_ident_doc]
-            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<::cached::async_sync::Mutex<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| ::cached::async_sync::Mutex::new(#cache_create));
-            // Cached function
-            #(#attributes)*
-            #visibility #signature_no_muts {
-                use cached::Cached;
-                let key = #key_convert_block;
-                {
-                    // check if the result is cached
-                    let mut cache = #cache_ident.lock().await;
-                    if let Some(result) = cache.cache_get(&key) {
-                        #return_cache_block
-                    }
-                }
-                #do_set_return_block
-            }
-            // Prime cached function
-            #[doc = #prime_fn_indent_doc]
-            #[allow(dead_code)]
-            #visibility #prime_sig {
-                use cached::Cached;
-                let key = #key_convert_block;
-                #prime_do_set_return_block
-            }
+    let expanded = quote! {
+        // Cached static
+        #[doc = #cache_ident_doc]
+        #cache_type
+        // Cached function
+        #(#attributes)*
+        #visibility #signature_no_muts {
+            use cached::Cached;
+            let key = #key_convert_block;
+            #do_set_return_block
         }
-    } else {
-        quote! {
-            // Cached static
-            #[doc = #cache_ident_doc]
-            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<std::sync::Mutex<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| std::sync::Mutex::new(#cache_create));
-            // Cached function
-            #(#attributes)*
-            #visibility #signature_no_muts {
-                use cached::Cached;
-                let key = #key_convert_block;
-                {
-                    // check if the result is cached
-                    let mut cache = #cache_ident.lock().unwrap();
-                    if let Some(result) = cache.cache_get(&key) {
-                        #return_cache_block
-                    }
-                }
-                #do_set_return_block
-            }
-            // Prime cached function
-            #[doc = #prime_fn_indent_doc]
-            #[allow(dead_code)]
-            #visibility #prime_sig {
-                use cached::Cached;
-                let key = #key_convert_block;
-                #prime_do_set_return_block
-            }
+        // Prime cached function
+        #[doc = #prime_fn_indent_doc]
+        #[allow(dead_code)]
+        #visibility #prime_sig {
+            use cached::Cached;
+            let key = #key_convert_block;
+            #prime_do_set_return_block
         }
     };
 
@@ -563,36 +375,7 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
     let asyncness = signature.asyncness;
 
     // pull out the names and types of the function inputs
-    let input_names = inputs
-        .iter()
-        .map(|input| match input {
-            FnArg::Receiver(_) => panic!("methods (functions taking 'self') are not supported"),
-            FnArg::Typed(pat_type) => {
-                // if you define arguments as mutable, e.g.
-                // #[once]
-                // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-                //     a += 1;
-                //     b += 1;
-                //     (a, b)
-                // }
-                // then we need to strip off the `mut` keyword from the
-                // variable identifiers so we can refer to arguments `a` and `b`
-                // instead of `mut a` and `mut b`
-                match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                }
-            }
-        })
-        .collect::<Vec<Box<Pat>>>();
+    let input_names = get_input_names(&inputs);
 
     // pull out the output type
     let output_ty = match &output {
@@ -602,69 +385,15 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
 
     let output_span = output_ty.span();
     let output_ts = TokenStream::from(output_ty.clone());
-    let output_parts = output_ts
-        .clone()
-        .into_iter()
-        .filter_map(|tt| match tt {
-            proc_macro::TokenTree::Ident(ident) => Some(ident.to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
+    let output_parts = get_output_parts(&output_ts);
     let output_string = output_parts.join("::");
     let output_type_display = output_ts.to_string().replace(' ', "");
 
-    // if `with_cached_flag = true`, then enforce that the return type
-    // is something wrapped in `Return`. Either `Return<T>` or the
-    // fully qualified `cached::Return<T>`
-    if args.with_cached_flag
-        && !output_string.contains("Return")
-        && !output_string.contains("cached::Return")
-    {
-        return syn::Error::new(
-            output_span,
-            format!(
-                "\nWhen specifying `with_cached_flag = true`, \
-                    the return type must be wrapped in `cached::Return<T>`. \n\
-                    The following return types are supported: \n\
-                    |    `cached::Return<T>`\n\
-                    |    `std::result::Result<cachedReturn<T>, E>`\n\
-                    |    `std::option::Option<cachedReturn<T>>`\n\
-                    Found type: {t}.",
-                t = output_type_display
-            ),
-        )
-        .to_compile_error()
-        .into();
+    if check_with_cache_flag(args.with_cached_flag, output_string) {
+        return with_cache_flag_error(output_span, output_type_display);
     }
 
-    // Find the type of the value to store.
-    // Normally it's the same as the return type of the functions, but
-    // for Options and Results it's the (first) inner type. So for
-    // Option<u32>, store u32, for Result<i32, String>, store i32, etc.
-    let cache_value_ty = match (&args.result, &args.option) {
-        (false, false) => output_ty,
-        (true, true) => panic!("the result and option attributes are mutually exclusive"),
-        _ => match output.clone() {
-            ReturnType::Default => {
-                panic!("function must return something for result or option attributes")
-            }
-            ReturnType::Type(_, ty) => {
-                if let Type::Path(typepath) = *ty {
-                    let segments = typepath.path.segments;
-                    if let PathArguments::AngleBracketed(brackets) =
-                        &segments.last().unwrap().arguments
-                    {
-                        let inner_ty = brackets.args.first().unwrap();
-                        quote! {#inner_ty}
-                    } else {
-                        panic!("function return type has no inner type")
-                    }
-                } else {
-                    panic!("function return type too complex")
-                }
-            }
-        },
-    };
+    let cache_value_ty = find_value_type(args.result, args.option, &output, output_ty);
 
     // make the cache identifier
     let cache_ident = match args.name {
@@ -699,16 +428,7 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
             } else {
                 quote! { return result.clone() }
             };
-            let return_cache_block = if let Some(time) = &args.time {
-                quote! {
-                    let (created_sec, result) = result;
-                    if now.duration_since(*created_sec).as_secs() < #time {
-                        #return_cache_block
-                    }
-                }
-            } else {
-                quote! { #return_cache_block }
-            };
+            let return_cache_block = gen_return_cache_block(args.time, return_cache_block);
             (set_cache_block, return_cache_block)
         }
         (true, false) => {
@@ -731,16 +451,7 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
             } else {
                 quote! { return Ok(result.clone()) }
             };
-            let return_cache_block = if let Some(time) = &args.time {
-                quote! {
-                    let (created_sec, result) = result;
-                    if now.duration_since(*created_sec).as_secs() < #time {
-                        #return_cache_block
-                    }
-                }
-            } else {
-                quote! { #return_cache_block }
-            };
+            let return_cache_block = gen_return_cache_block(args.time, return_cache_block);
             (set_cache_block, return_cache_block)
         }
         (false, true) => {
@@ -763,133 +474,98 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
             } else {
                 quote! { return Some(result.clone()) }
             };
-            let return_cache_block = if let Some(time) = &args.time {
-                quote! {
-                    let (created_sec, result) = result;
-                    if now.duration_since(*created_sec).as_secs() < #time {
-                        #return_cache_block
-                    }
-                }
-            } else {
-                quote! { #return_cache_block }
-            };
+            let return_cache_block = gen_return_cache_block(args.time, return_cache_block);
             (set_cache_block, return_cache_block)
         }
         _ => panic!("the result and option attributes are mutually exclusive"),
     };
 
-    let do_set_return_block = if asyncness.is_some() {
-        if args.sync_writes {
-            quote! {
-                // try to get a write lock first
-                let mut cached = #cache_ident.write().await;
-                if let Some(result) = &*cached {
-                    #return_cache_block
-                }
+    let set_cache_and_return = quote! {
+        #set_cache_block
+        result
+    };
+    let r_lock;
+    let w_lock;
+    let function_call;
+    let cache_type;
+    if asyncness.is_some() {
+        w_lock = quote! {
+            // try to get a write lock
+            let mut cached = #cache_ident.write().await;
+        };
 
-                // run the function and cache the result
-                async fn inner(#inputs) #output #body;
-                let result = inner(#(#input_names),*).await;
-                #set_cache_block
-                result
-            }
-        } else {
-            quote! {
-                // run the function and cache the result
-                async fn inner(#inputs) #output #body;
-                let result = inner(#(#input_names),*).await;
-                let mut cached = #cache_ident.write().await;
-                #set_cache_block
-                result
-            }
-        }
-    } else if args.sync_writes {
-        quote! {
-            // try to get a write lock first
+        r_lock = quote! {
+            // try to get a read lock
+            let mut cached = #cache_ident.read().await;
+        };
+
+        function_call = quote! {
+            // run the function and cache the result
+            async fn inner(#inputs) #output #body;
+            let result = inner(#(#input_names),*).await;
+        };
+
+        cache_type = quote! {
+            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<::cached::async_sync::RwLock<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| ::cached::async_sync::RwLock::new(#cache_create));
+        };
+    } else {
+        w_lock = quote! {
+            // try to get a lock first
             let mut cached = #cache_ident.write().unwrap();
+        };
+
+        r_lock = quote! {
+            // try to get a read lock
+            let mut cached = #cache_ident.read().unwrap();
+        };
+
+        function_call = quote! {
+            // run the function and cache the result
+            fn inner(#inputs) #output #body;
+            let result = inner(#(#input_names),*);
+        };
+
+        cache_type = quote! {
+            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<std::sync::RwLock<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| std::sync::RwLock::new(#cache_create));
+        };
+    }
+
+    let prime_do_set_return_block = quote! {
+        #w_lock
+        #function_call
+        #set_cache_and_return
+    };
+
+    let return_cache_block = quote! {
+        {
+            #r_lock
             if let Some(result) = &*cached {
                 #return_cache_block
             }
-
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            #set_cache_block
-            result
-        }
-    } else {
-        quote! {
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            let mut cached = #cache_ident.write().unwrap();
-            #set_cache_block
-            result
         }
     };
 
-    // if you define arguments as mutable, e.g.
-    // #[once]
-    // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-    //     a += 1;
-    //     b += 1;
-    //     (a, b)
-    // }
-    // then we want the `mut` keywords present on the "inner" function
-    // that wraps your actual block of code.
-    // If the `mut`s are also on the outer method, then you'll
-    // get compiler warnings about your arguments not needing to be `mut`
-    // when they really do need to be.
-    let mut signature_no_muts = signature;
-    let mut sig_inputs = Punctuated::new();
-    for inp in &signature_no_muts.inputs {
-        let item = match inp {
-            FnArg::Receiver(_) => inp.clone(),
-            FnArg::Typed(pat_type) => {
-                let mut pt = pat_type.clone();
-                let pat = match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                };
-                pt.pat = pat;
-                FnArg::Typed(pt)
-            }
-        };
-        sig_inputs.push(item);
-    }
-    signature_no_muts.inputs = sig_inputs;
+    let do_set_return_block = if args.sync_writes {
+        quote! {
+            #return_cache_block
+            #w_lock
+            #function_call
+            #set_cache_and_return
+        }
+    } else {
+        quote! {
+            #return_cache_block
+            #function_call
+            #w_lock
+            #set_cache_and_return
+        }
+    };
+
+    let signature_no_muts = get_mut_signature(signature);
 
     let prime_fn_ident = Ident::new(&format!("{}_prime_cache", &fn_ident), fn_ident.span());
     let mut prime_sig = signature_no_muts.clone();
     prime_sig.ident = prime_fn_ident;
-
-    let prime_do_set_return_block = if asyncness.is_some() {
-        quote! {
-            // run the function and cache the result
-            async fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*).await;
-            let mut cached = #cache_ident.write().await;
-            #set_cache_block
-            result
-        }
-    } else {
-        quote! {
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            let mut cached = #cache_ident.write().unwrap();
-            #set_cache_block
-            result
-        }
-    };
 
     // make cached static, cached function and prime cached function doc comments
     let cache_ident_doc = format!("Cached static for the [`{}`] function.", fn_ident);
@@ -898,66 +574,25 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
         "This is a cached function that uses the [`{}`] cached static.",
         cache_ident
     );
-    if attributes.iter().any(|attr| attr.path.is_ident("doc")) {
-        attributes.push(parse_quote! { #[doc = ""] });
-        attributes.push(parse_quote! { #[doc = "# Caching"] });
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    } else {
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    }
+    fill_in_attributes(&mut attributes, cache_fn_doc_extra);
 
     // put it all together
-    let expanded = if asyncness.is_some() {
-        quote! {
-            // Cached static
-            #[doc = #cache_ident_doc]
-            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<::cached::async_sync::RwLock<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| ::cached::async_sync::RwLock::new(#cache_create));
-            // Cached function
-            #(#attributes)*
-            #visibility #signature_no_muts {
-                let now = ::cached::instant::Instant::now();
-                {
-                    // check if the result is cached
-                    let mut cached = #cache_ident.read().await;
-                    if let Some(result) = &*cached {
-                        #return_cache_block
-                    }
-                }
-                #do_set_return_block
-            }
-            // Prime cached function
-            #[doc = #prime_fn_indent_doc]
-            #[allow(dead_code)]
-            #visibility #prime_sig {
-                let now = ::cached::instant::Instant::now();
-                #prime_do_set_return_block
-            }
+    let expanded = quote! {
+        // Cached static
+        #[doc = #cache_ident_doc]
+        #cache_type
+        // Cached function
+        #(#attributes)*
+        #visibility #signature_no_muts {
+            let now = ::cached::instant::Instant::now();
+            #do_set_return_block
         }
-    } else {
-        quote! {
-            // Cached static
-            #[doc = #cache_ident_doc]
-            #visibility static #cache_ident: ::cached::once_cell::sync::Lazy<std::sync::RwLock<#cache_ty>> = ::cached::once_cell::sync::Lazy::new(|| std::sync::RwLock::new(#cache_create));
-            // Cached function
-            #(#attributes)*
-            #visibility #signature_no_muts {
-                let now = ::cached::instant::Instant::now();
-                {
-                    // check if the result is cached
-                    let mut cached = #cache_ident.read().unwrap();
-                    if let Some(result) = &*cached {
-                        #return_cache_block
-                    }
-                }
-                #do_set_return_block
-            }
-            // Prime cached function
-            #[doc = #prime_fn_indent_doc]
-            #[allow(dead_code)]
-            #visibility #prime_sig {
-                let now = ::cached::instant::Instant::now();
-                #prime_do_set_return_block
-            }
+        // Prime cached function
+        #[doc = #prime_fn_indent_doc]
+        #[allow(dead_code)]
+        #visibility #prime_sig {
+            let now = ::cached::instant::Instant::now();
+            #prime_do_set_return_block
         }
     };
 
@@ -1000,7 +635,7 @@ struct IOMacroArgs {
 /// - `cache_prefix_block`: (optional, string expr) specify an expression used to create the string used as a
 ///   prefix for all cache keys of this function, e.g. `cache_prefix_block = r##"{ "my_prefix" }"##`.
 ///   When not specified, the cache prefix will be constructed from the name of the function. This
-///   could result in unexpected conflicts between io_cached-functions of the same name so it's
+///   could result in unexpected conflicts between io_cached-functions of the same name, so it's
 ///   recommended that you specify a prefix you're sure will be unique.
 /// - `create`: (optional, string expr) specify an expression used to create a new cache store, e.g. `create = r##"{ CacheType::new() }"##`.
 /// - `key`: (optional, string type) specify what type to use for the cache key, e.g. `type = "TimedCached<u32, u32>"`.
@@ -1038,45 +673,9 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
     let output = signature.output.clone();
     let asyncness = signature.asyncness;
 
-    // pull out the names and types of the function inputs
-    let input_tys = inputs
-        .iter()
-        .map(|input| match input {
-            FnArg::Receiver(_) => panic!("methods (functions taking 'self') are not supported"),
-            FnArg::Typed(pat_type) => pat_type.ty.clone(),
-        })
-        .collect::<Vec<Box<Type>>>();
+    let input_tys = get_input_types(&inputs);
 
-    let input_names = inputs
-        .iter()
-        .map(|input| match input {
-            FnArg::Receiver(_) => panic!("methods (functions taking 'self') are not supported"),
-            FnArg::Typed(pat_type) => {
-                // if you define arguments as mutable, e.g.
-                // #[cached]
-                // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-                //     a += 1;
-                //     b += 1;
-                //     (a, b)
-                // }
-                // then we need to strip off the `mut` keyword from the
-                // variable identifiers so we can refer to arguments `a` and `b`
-                // instead of `mut a` and `mut b`
-                match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                }
-            }
-        })
-        .collect::<Vec<Box<Pat>>>();
+    let input_names = get_input_names(&inputs);
 
     // pull out the output type
     let output_ty = match &output {
@@ -1086,14 +685,7 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
 
     let output_span = output_ty.span();
     let output_ts = TokenStream::from(output_ty);
-    let output_parts = output_ts
-        .clone()
-        .into_iter()
-        .filter_map(|tt| match tt {
-            proc_macro::TokenTree::Ident(ident) => Some(ident.to_string()),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
+    let output_parts = get_output_parts(&output_ts);
     let output_string = output_parts.join("::");
     let output_type_display = output_ts.to_string().replace(' ', "");
 
@@ -1172,33 +764,17 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
 
     // make the cache identifier
     let cache_ident = match args.name {
-        Some(name) => Ident::new(&name, fn_ident.span()),
+        Some(ref name) => Ident::new(name, fn_ident.span()),
         None => Ident::new(&fn_ident.to_string().to_uppercase(), fn_ident.span()),
     };
 
-    // make the cache key type and block that converts the inputs into the key type
-    let (cache_key_ty, key_convert_block) = match (&args.key, &args.convert, &args.cache_type) {
-        (Some(key_str), Some(convert_str), _) => {
-            let cache_key_ty = parse_str::<Type>(key_str).expect("unable to parse cache key type");
-
-            let key_convert_block =
-                parse_str::<Block>(convert_str).expect("unable to parse key convert block");
-
-            (quote! {#cache_key_ty}, quote! {#key_convert_block})
-        }
-        (None, Some(convert_str), Some(_)) => {
-            let key_convert_block =
-                parse_str::<Block>(convert_str).expect("unable to parse key convert block");
-
-            (quote! {}, quote! {#key_convert_block})
-        }
-        (None, None, _) => (
-            quote! {(#(#input_tys),*)},
-            quote! {(#(#input_names.clone()),*)},
-        ),
-        (Some(_), None, _) => panic!("key requires convert to be set"),
-        (None, Some(_), None) => panic!("convert requires key or type to be set"),
-    };
+    let (cache_key_ty, key_convert_block) = make_cache_key_type(
+        &args.key,
+        &args.convert,
+        &args.cache_type,
+        input_tys,
+        &input_names,
+    );
 
     // make the cache type and create statement
     let (cache_ty, cache_create) = match (
@@ -1368,69 +944,12 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
         }
     };
 
-    // if you define arguments as mutable, e.g.
-    // #[cached]
-    // fn mutable_args(mut a: i32, mut b: i32) -> (i32, i32) {
-    //     a += 1;
-    //     b += 1;
-    //     (a, b)
-    // }
-    // then we want the `mut` keywords present on the "inner" function
-    // that wraps your actual block of code.
-    // If the `mut`s are also on the outer method, then you'll
-    // get compiler warnings about your arguments not needing to be `mut`
-    // when they really do need to be.
-    let mut signature_no_muts = signature;
-    let mut sig_inputs = Punctuated::new();
-    for inp in &signature_no_muts.inputs {
-        let item = match inp {
-            FnArg::Receiver(_) => inp.clone(),
-            FnArg::Typed(pat_type) => {
-                let mut pt = pat_type.clone();
-                let pat = match &pat_type.pat.deref() {
-                    Pat::Ident(pat_ident) => {
-                        if pat_ident.mutability.is_some() {
-                            let mut p = pat_ident.clone();
-                            p.mutability = None;
-                            Box::new(Pat::Ident(p))
-                        } else {
-                            Box::new(Pat::Ident(pat_ident.clone()))
-                        }
-                    }
-                    _ => pat_type.pat.clone(),
-                };
-                pt.pat = pat;
-                FnArg::Typed(pt)
-            }
-        };
-        sig_inputs.push(item);
-    }
-    signature_no_muts.inputs = sig_inputs;
+    let signature_no_muts = get_mut_signature(signature);
 
     // create a signature for the cache-priming function
     let prime_fn_ident = Ident::new(&format!("{}_prime_cache", &fn_ident), fn_ident.span());
     let mut prime_sig = signature_no_muts.clone();
     prime_sig.ident = prime_fn_ident;
-
-    let prime_do_set_return_block = if asyncness.is_some() {
-        quote! {
-            // run the function and cache the result
-            async fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*).await;
-            let cache = &#cache_ident.get().await;
-            #set_cache_block
-            result
-        }
-    } else {
-        quote! {
-            // run the function and cache the result
-            fn inner(#inputs) #output #body;
-            let result = inner(#(#input_names),*);
-            let cache = &#cache_ident;
-            #set_cache_block
-            result
-        }
-    };
 
     // make cached static, cached function and prime cached function doc comments
     let cache_ident_doc = format!("Cached static for the [`{}`] function.", fn_ident);
@@ -1439,13 +958,7 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
         "This is a cached function that uses the [`{}`] cached static.",
         cache_ident
     );
-    if attributes.iter().any(|attr| attr.path.is_ident("doc")) {
-        attributes.push(parse_quote! { #[doc = ""] });
-        attributes.push(parse_quote! { #[doc = "# Caching"] });
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    } else {
-        attributes.push(parse_quote! { #[doc = #cache_fn_doc_extra] });
-    }
+    fill_in_attributes(&mut attributes, cache_fn_doc_extra);
 
     // put it all together
     let expanded = if asyncness.is_some() {
@@ -1475,7 +988,7 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
             #visibility #prime_sig {
                 use cached::IOCachedAsync;
                 let key = #key_convert_block;
-                #prime_do_set_return_block
+                #do_set_return_block
             }
         }
     } else {
@@ -1503,7 +1016,7 @@ pub fn io_cached(args: TokenStream, input: TokenStream) -> TokenStream {
             #visibility #prime_sig {
                 use cached::IOCached;
                 let key = #key_convert_block;
-                #prime_do_set_return_block
+                #do_set_return_block
             }
         }
     };
